(untested) Signal overlapped from caller in LUsb0_ControlTransfer
This is an attempt to deal with libusb 1.0.24 or later which always uses
asynchronous control transfers. It will pass an OVERLAPPED structure and
use this afterwards to check that the transfer has finished.

It will still be forced into a synchronous transfer here, but by
signalling the finished transfer the caller will not be reporting
timeouts for actual finished transfers.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>